@@ -22,8 +22,8 @@ class NearbySightsFinder {
     mocks.forEach((element) {
       allSights.add(Sight(
           name: element[0],
-          lat: element[1],
-          lon: element[2],
+          lat: double.parse(element[1]),
+          lon: double.parse(element[2]),
           url: element[3],
           details: element[4],
           type: element[5]));
@@ -45,11 +45,11 @@ class NearbySightsFinder {
     await Geolocator.checkPermission();
   } */
 
-  bool isPointsNear(List<double> checkPoint, List<double> centerPoint, double km) {
+  bool isPointsNear(List<double?> checkPoint, List<double> centerPoint, double km) {
     var ky = 40000 / 360;
     var kx = cos(pi * centerPoint[0] / 180.0) * ky;
-    var dx = (centerPoint[1] - checkPoint[1]).abs() * kx;
-    var dy = (centerPoint[0] - checkPoint[0]) * ky;
+    var dx = (centerPoint[1] - checkPoint[1]!).abs() * kx;
+    var dy = (centerPoint[0] - checkPoint[0]!) * ky;
     return sqrt(dx * dx + dy * dy) <= km;
   }
  
