@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/mocks.dart';
+import 'package:places/ui/screen/add_sight_screen.dart';
 import 'package:places/ui/screen/filter_screen.dart';
 import 'package:places/ui/screen/res/themes.dart';
+import 'package:places/ui/screen/search_screen.dart';
 import 'package:places/ui/screen/settings_screen.dart';
 import 'package:places/ui/screen/sight_details.dart';
 import 'package:places/ui/screen/sight_list_screen.dart';
@@ -32,10 +34,14 @@ class _AppState extends State<App> {
       builder: (context, themeProvider, child) {
         return MaterialApp(
         title: 'First App',
+      //  home: SearchScreen(),
+       // home: AddSightScreenCategory(setCategory: (value){}),
+       // home: AddSightScreenCategory(setCategory: (value){} ),
+       // home: AddSightScreen(),
        // home: SettingsScreen(),
         // home: FilterScreen(),
-         home: VisitingScreen(),
-        // home: SightListScreen(title: "Список интересных мест",),
+        // home: VisitingScreen(),
+         home: SightListScreen(title: "Список интересных мест",),
         // home: SightDetails(sight: Sight(name: mocks[0][0], lat: mocks[0][1], lon: mocks[0][2], url: mocks[0][3], details: mocks[0][4], type: mocks[0][5]),),
         //theme: globals.isDarkMode ? darkTheme : lightTheme,
         theme: themeProvider.getTheme,
